eduConnect/requirement.txt
Django==4.0.0 
djangorestframework==3.13.1 
psycopg2-binary==2.9.3 
django-filter==21.1 
python-dotenv==0.19.2